Add cgroup support in the process view
Add a new 'Control Group' column to the process view tab.
The format for the column is:

<path name> (controller name), <path name> (controller name)...

Processes that share the same path name across controllers
are colesced. For example if a process is in the /foo cgroup
for both the memory and cpu controllers, it would display as:
/foo (memory,cpu), ...

Taken from GSM commits:
2d33adcbc4347c112d57082956b4e199ff7132db Add cgroup support in the process view
645d38a218fedcbf9c7b674740bf213e02933820 Optimize cgroup updating

From: Jason Baron <jbaron@redhat.com>