Count bandwidth usage in bytes
In the Network History section of mate-system-monitor, currently, the
"Total Recieved" and "Total Sent" values are calculated in bits if the
"Show network speed in bits" option is enabled in Preferences. With this
commit, those are always measured in bytes, even when the "Receiving"
and "Sending" values are set to be measured in bits.