interface: Allows Delete key to end processes
The delete key now can end selected processes as Ctrl-E can while
the menu entry displaying the Ctrl+E shortcut remains. The purpose of
this is to allow convenience for those who are switching over from
gnome-system-monitor or other process managers which allow the
delete key to End selected processes. This was accomplished by
adding an accelerator item with the delete key as the accelerator
and putting the menu entry of the EndProcess_Del on top of the
EndProcess menu entry so the latter shows over the former.

Issue #101